+++ config/api/category/fields.php
@@ -29,6 +29,9 @@ return [
         'title' => 'category/fields.title-resource_type_id',
         'description' => 'category/fields.description-resource_type_id',
         'type' => 'string',
+        'validation' => [
+            'length' => 10
+        ],
         'required' => true
     ]
 ];