+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 The complete changelog for the Costs to Expect REST API, follows the format defined at https://keepachangelog.com/en/1.0.0/
 
 ## [v1.23.0] - 2019-09-xx
-
+## Changed
+- We have added string length validation for hashed id values; all should be ten characters.
 
 ## [v1.22.1] - 2019-09-01
 ### Changed